+++ rmq_metadata.py
@@ -511,9 +511,10 @@ def _convert_data(rmq, log, cfg, queue, body, r_key):
         log.log_info("Finished processing of: %s" % (f_filename))
 
     else:
-        log.log_err("All extractions failed on: %s" % (f_filename))
+        log.log_err("Insert or extractions failed on: %s" % (f_filename))
         log.log_info("Body of message being saved to a file - see below")
-        non_proc_msg(rmq, log, cfg, body, "All extractions failed", r_key)
+        non_proc_msg(rmq, log, cfg, body,
+                     "All extractions or Mongo insertion failure", r_key)
         os.remove(f_name)
         log.log_info("Cleanup of temporary files completed.")
         log.log_info("Finished processing of: %s" % (f_filename))